requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
django-sslserver==0.20
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
pycrypto==2.6.1
PyJWT==1.7.1
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
tablib==0.13.0
urllib3==1.24.3
xlrd==1.2.0
xlwt==1.3.0
xpinyin==0.5.6
